Add files via upload
"Shopperz" is a dynamic e-commerce platform designed to provide a seamless shopping experience, similar to industry giants like Amazon and Flipkart. This project leverages modern front-end technologies to create a visually appealing and user-friendly interface. Key features include multiple slides showcasing a variety of products with an interactive and responsive design ensuring smooth transitions, a powerful search bar with auto-suggestions and filters to help users quickly find products, and a homepage displaying a diverse array of products in a categorized and organized layout for easy navigation. The goal of Shopperz is to create an engaging and efficient online shopping experience, focusing on user-friendly navigation and fast, accurate search results, demonstrating proficiency in front-end development by building a feature-rich e-commerce site.